Cursor_template/requirements.txt
streamlit>=1.28.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-extras>=0.3.0
streamlit-ace>=0.1.1
streamlit-authenticator>=0.2.3
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
sqlite3
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
xlrd>=2.0.0
